fix(api): use u64 for session IDs to prevent overflow after 4B connections
This commit resolves finding M-1 from the security audit by changing the
session API to use u64 for connection IDs, preventing potential overflow
and ID collision after 4 billion connections.

**Changes**:
- SessionInfo.id: Changed from u32 to u64 with documentation
- get_sessions() handler: Removed unsafe u32 cast, use full u64
- Test assertions: Updated to expect u64 instead of u32

**Vulnerability Details** (CWE-681):
The session API previously converted u64 connection IDs to u32 without bounds
checking. After 4,294,967,296 connections (u32::MAX), the conversion would
truncate, causing session ID collision in API responses.

**Fix**:
```rust
// Before (unsafe conversion):
pub id: u32,
// ...
id: r.connection_id as u32,

// After (full precision):
pub id: u64,
// ...
id: r.connection_id,
```

**Security Impact**:
- Prevents session ID collision in API responses
- Maintains 1:1 mapping between internal connection IDs and API session IDs
- Low probability issue (requires 4B+ connections) but now eliminated
- Does not affect actual session tracking (always used u64 internally)

**Compatibility**:
- API response format change: session IDs now return u64 instead of u32
- Backward compatible: u32 values are valid u64 values
- No database schema changes (API-only)

All 252 tests passing.

NIST Controls: AU-3 (Content of Audit Records), AC-10 (Concurrent Session Control)

+++ crates/tacacs-server/src/api/handlers.rs
@@ -259,7 +259,7 @@ async fn get_sessions(State(state): State<Arc<ApiState>>) -> impl IntoResponse {
     let sessions: Vec<SessionInfo> = records
         .iter()
         .map(|r| SessionInfo {
-            id: r.connection_id as u32, // Use connection_id as the session identifier
+            id: r.connection_id, // Use connection_id as the session identifier (u64)
             peer_addr: r.peer_addr.to_string(),
             username: r.username.clone(),
             start_time: r
@@ -852,7 +852,7 @@ mod tests {
 
         assert_eq!(sessions_response.total, 1);
         assert_eq!(sessions_response.sessions.len(), 1);
-        assert_eq!(sessions_response.sessions[0].id, conn_id as u32);
+        assert_eq!(sessions_response.sessions[0].id, conn_id);
         assert_eq!(
             sessions_response.sessions[0].peer_addr,
             "192.168.1.100:54321"
@@ -1046,10 +1046,10 @@ mod tests {
         assert_eq!(sessions_response.sessions.len(), 3);
 
         // Verify session details
-        let conn_ids: Vec<u32> = sessions_response.sessions.iter().map(|s| s.id).collect();
-        assert!(conn_ids.contains(&(conn1 as u32)));
-        assert!(conn_ids.contains(&(conn2 as u32)));
-        assert!(conn_ids.contains(&(conn3 as u32)));
+        let conn_ids: Vec<u64> = sessions_response.sessions.iter().map(|s| s.id).collect();
+        assert!(conn_ids.contains(&conn1));
+        assert!(conn_ids.contains(&conn2));
+        assert!(conn_ids.contains(&conn3));
 
         // Cleanup
         registry.unregister_connection(conn1).await;
